+++ src/main/java/br/com/kg/estoque/domain/material/MaterialCustomRepository.java
@@ -68,7 +68,7 @@ public class MaterialCustomRepository{
 		var query = em.createQuery(jpql.toString(), Long.class);
 		query.setParameter("searchValue", "%"+sSearch.toLowerCase()+"%");
 		
-		return (Long) query.getSingleResult();
+		return query.getSingleResult();
 		
 	}
